Refactor: 알림창 컴포넌트 분리 및 이미지 적용
