--- src/app/components/common/Header/Header.module.scss
@@ -1,265 +1,0 @@
-.wrapper {
-  width: 100%;
-  background-color: #ffffff;
-}
-.contentWrapper {
-  width: 1440px;
-  margin: 0 auto;
-  display: grid;
-  grid-template: 'logo searchBar buttonWrapper'/ 7.5rem 28.125rem;
-  padding: 15px 200px;
-  gap: 1.25rem;
-  place-items: center;
-}
-
-.logo {
-  grid-area: logo;
-}
-
-.searchBarWrapper {
-  grid-area: searchBar;
-  width: 28.125rem;
-  height: 2.5rem;
-  position: relative;
-}
-
-.searchButton {
-  position: absolute;
-  left: 0.75rem;
-  top: 0.625rem;
-}
-
-.searchBar {
-  width: 100%;
-  height: 100%;
-  border-radius: 0.625rem;
-  padding: 0.625rem;
-  padding-left: 2.625rem;
-  background-color: #f2f2f3;
-  color: #a4a1aa;
-  outline: none;
-  border: 0;
-  font-weight: 400;
-  font-size: 0.875rem;
-  line-height: 1.375rem;
-}
-
-.buttonWrapper {
-  grid-area: buttonWrapper;
-  width: 100%;
-  display: flex;
-  gap: 2.5rem;
-  align-items: center;
-  justify-content: flex-end;
-}
-
-@mixin button {
-  font-weight: 700;
-  font-size: 1rem;
-  line-height: 1.25rem;
-  color: #111322;
-  cursor: pointer;
-
-  @media (max-width: 736px) {
-    font-size: 0.875rem;
-    line-height: 1.25rem;
-  }
-}
-
-.button {
-  @include button;
-}
-
-.notificationWindow {
-  position: absolute;
-  width: 368px;
-  background-color: #ffebe7;
-  border: 1px solid #cbc9cf;
-  border-radius: 10px;
-  box-shadow: 0px 2px 8px 0px rgba(120, 116, 134, 0.25);
-  height: 419px;
-  padding: 24px 20px;
-  top: 40px;
-  right: 5px;
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-  cursor: initial;
-}
-
-.notification {
-  @include button;
-  position: relative;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 1.5rem;
-  height: 1.5rem;
-  cursor: pointer;
-}
-
-.notificationTitleWrapper {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.notificationTitle {
-  font-weight: 700;
-  font-size: 20px;
-  line-height: 24px;
-  color: #111322;
-}
-
-.notificationClose {
-  display: none;
-  cursor: pointer;
-}
-
-.notificationList {
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-  overflow: auto;
-  height: 335px;
-
-  &::-webkit-scrollbar {
-    background-color: transparent;
-    width: 3px;
-  }
-
-  &::-webkit-scrollbar-thumb {
-    background-color: #888; /* 스크롤바 색상을 지정 */
-  }
-}
-
-.notificationEmpty {
-  margin: auto;
-}
-
-@mixin notificationWrapper {
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-  padding: 16px 12px;
-  width: 100%;
-  background-color: #ffffff;
-  border: 1px solid #e5e4e7;
-  border-radius: 5px;
-  cursor: pointer;
-}
-
-.notificationReadWrapper {
-  @include notificationWrapper;
-  opacity: 0.7;
-}
-
-.notificationWrapper {
-  @include notificationWrapper;
-}
-
-@mixin fin {
-  width: 5px;
-  height: 5px;
-  border-radius: 5px;
-}
-
-.notificationRedFin {
-  @include fin;
-  background-color: #ff4040;
-}
-
-.notificationBlueFin {
-  @include fin;
-  background-color: #0080ff;
-}
-
-.notificationText {
-  font-weight: 400;
-  font-size: 14px;
-  line-height: 22px;
-}
-
-.notificationBlueStatus {
-  color: #0080ff;
-}
-
-.notificationRedStatus {
-  color: #ff4040;
-}
-
-.notificationDate {
-  font-weight: 400;
-  font-size: 12px;
-  line-height: 16px;
-  color: #a4a1aa;
-}
-
-@media (max-width: 1440px) {
-  .contentWrapper {
-    width: 1040px;
-    padding: 15px 30px;
-    margin: 0 auto;
-  }
-}
-
-@media (max-width: 1040px) {
-  .contentWrapper {
-    width: 100%;
-    padding: 15px 30px;
-  }
-}
-
-/* tablet */
-@media (max-width: 1023px) {
-  .contentWrapper {
-    grid-template: 'logo searchBar buttonWrapper'/ 7.5rem 21.5rem;
-    padding: 15px 30px;
-    gap: 1rem;
-  }
-  .searchBarWrapper {
-    width: 21.5rem;
-  }
-  .buttonWrapper {
-    gap: 0.75rem;
-  }
-}
-
-/* mobile */
-@media (max-width: 767px) {
-  .contentWrapper {
-    min-width: none;
-    width: 100vw;
-    grid-template:
-      'logo buttonWrapper'
-      'searchBar searchBar'
-      /84px;
-  }
-  .searchBarWrapper {
-    width: 100%;
-  }
-  .logo {
-    width: 84px;
-    height: 30px;
-  }
-  .notification {
-    position: initial;
-  }
-  .notificationWindow {
-    top: 0;
-    left: 0;
-    width: 100vw;
-    height: 100vh;
-    padding: 40px 20px;
-    border: 0;
-    border-radius: 0;
-    box-shadow: none;
-  }
-  .notificationClose {
-    display: block;
-  }
-  .notificationList {
-    height: 100%;
-    width: 100%;
-  }
-}
