+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import Header from './components/common/Header/Header';
+import Header from '../components/common/Header/Header';
 
 export default function Home() {
   const data = [
@@ -49,7 +49,7 @@ export default function Home() {
     },
   ];
   return (
-    <div style={{ width: '100vw', height: '100vh' }}>
+    <div>
       <Header memberType="worker" notificationListData={data} />
     </div>
   );
