+++ src/app/utils/constants.ts
@@ -1,0 +1,7 @@
+const BASE_PATH = '/assets';
+
+export const CLOSE_ICON = `${BASE_PATH}/close_icon.svg`;
+export const HEADER_LOGO = `${BASE_PATH}/header_logo.svg`;
+export const NOTIFICATION_ACTIVE = `${BASE_PATH}/notification_active.svg`;
+export const NOTIFICATION_INACTIVE = `${BASE_PATH}/notification_inactive.svg`;
+export const SEARCH_ICON = `${BASE_PATH}/search_icon.svg`;
