+++ src/components/common/Header/Header.module.scss
@@ -1,0 +1,140 @@
+@import '@/styles/_index';
+
+.wrapper {
+  width: 100%;
+  background-color: $white;
+}
+.contentWrapper {
+  box-sizing: border-box;
+  width: 144rem;
+  margin: 0 auto;
+  display: grid;
+  grid-template: 'logo searchBar buttonWrapper'/ 12rem 45rem;
+  padding: 1.5rem 20rem;
+  gap: 2rem;
+  place-items: center;
+}
+
+.logo {
+  grid-area: logo;
+}
+
+.searchBarWrapper {
+  grid-area: searchBar;
+  width: 45rem;
+  height: 4rem;
+  position: relative;
+}
+
+.searchButton {
+  position: absolute;
+  left: 1rem;
+  top: 1rem;
+}
+
+.searchBar {
+  box-sizing: border-box;
+  width: 100%;
+  height: 100%;
+  border-radius: 1rem;
+  padding: 1rem;
+  padding-left: 4.2rem;
+  background-color: $gray-200;
+  color: $gray-500;
+  outline: none;
+  border: 0;
+  @include font-normal;
+  font-size: 1.4rem;
+  line-height: 2.2rem;
+}
+
+@mixin buttonWrapper {
+  grid-area: buttonWrapper;
+  width: 100%;
+  display: flex;
+  gap: 4rem;
+  align-items: center;
+  justify-content: flex-end;
+}
+
+.buttonWrapper {
+  @include buttonWrapper;
+}
+
+.buttonWrapperSmall {
+  @include buttonWrapper;
+}
+
+@mixin button {
+  @include font-bold;
+  font-size: 1.6rem;
+  line-height: 2rem;
+  color: $black;
+  cursor: pointer;
+
+  @media (max-width: 767px) {
+    font-size: 1.4rem;
+    line-height: 2rem;
+  }
+}
+
+.button {
+  @include button;
+}
+
+@media (max-width: 1440px) {
+  .contentWrapper {
+    width: 104rem;
+    padding: 1.5rem 3rem;
+    margin: 0 auto;
+  }
+}
+
+@media (max-width: 1040px) {
+  .contentWrapper {
+    width: 100%;
+    padding: 1.5rem 3rem;
+  }
+}
+
+/* tablet */
+@media (max-width: 1023px) {
+  .contentWrapper {
+    grid-template: 'logo searchBar buttonWrapper'/ 12rem 34.4rem;
+    padding: 1.5rem 3rem;
+    gap: 1.6rem;
+  }
+  .searchBarWrapper {
+    width: 34.4rem;
+  }
+  .buttonWrapper {
+    gap: 1.2rem;
+  }
+}
+
+/* mobile */
+@media (max-width: 767px) {
+  .contentWrapper {
+    min-width: none;
+    width: 100%;
+    grid-template:
+      'logo buttonWrapper'
+      'searchBar searchBar'
+      /8.4rem;
+  }
+  .searchBarWrapper {
+    width: 100%;
+  }
+  .logo {
+    width: 8.4rem;
+    height: 3rem;
+  }
+
+  .buttonWrapper {
+    gap: 1.6rem;
+  }
+
+  .buttonWrapperSmall {
+    gap: 1.6rem;
+  }
+}
