+++ src/components/common/Header/Notification.module.scss
@@ -1,0 +1,163 @@
+@import '@/styles/_index';
+
+.notificationWindow {
+  box-sizing: border-box;
+  position: absolute;
+  width: 36.8rem;
+  background-color: #ffebe7;
+  border: 0.1rem solid $gray-400;
+  border-radius: 1rem;
+  box-shadow: 0rem 0.2rem 0.8rem 0rem rgba(120, 116, 134, 0.25);
+  height: 41.9rem;
+  padding: 2.4rem 2rem;
+  top: 4rem;
+  right: 0.5rem;
+  display: flex;
+  flex-direction: column;
+  gap: 1.6rem;
+  cursor: initial;
+}
+
+@mixin button {
+  @include font-bold;
+  font-size: 1.6rem;
+  line-height: 2rem;
+  color: $black;
+  cursor: pointer;
+
+  @media (max-width: 767px) {
+    font-size: 1.4rem;
+    line-height: 2rem;
+  }
+}
+.notification {
+  @include button;
+  position: relative;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 2.4rem;
+  height: 2.4rem;
+  cursor: pointer;
+}
+
+.notificationTitleWrapper {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+
+.notificationTitle {
+  @include font-bold;
+  font-size: 2rem;
+  line-height: 2.4rem;
+  color: $black;
+}
+
+.notificationClose {
+  display: none;
+  cursor: pointer;
+}
+
+.notificationList {
+  display: flex;
+  flex-direction: column;
+  gap: 0.8rem;
+  overflow: auto;
+  height: 33.5rem;
+
+  &::-webkit-scrollbar {
+    background-color: transparent;
+    width: 0.3rem;
+  }
+
+  &::-webkit-scrollbar-thumb {
+    background-color: #888;
+  }
+}
+
+.notificationEmpty {
+  margin: auto;
+}
+
+@mixin notificationWrapper {
+  box-sizing: border-box;
+  display: flex;
+  flex-direction: column;
+  gap: 0.4rem;
+  padding: 1.6rem 1.2rem;
+  width: 100%;
+  background-color: $white;
+  border: 0.1rem solid $gray-300;
+  border-radius: 0.5rem;
+  cursor: pointer;
+}
+
+.notificationReadWrapper {
+  @include notificationWrapper;
+  opacity: 0.7;
+}
+
+.notificationWrapper {
+  @include notificationWrapper;
+}
+
+@mixin fin {
+  width: 0.5rem;
+  height: 0.5rem;
+  border-radius: 0.5rem;
+}
+
+.notificationRedFin {
+  @include fin;
+  background-color: $red-400;
+}
+
+.notificationBlueFin {
+  @include fin;
+  background-color: $blue-200;
+}
+
+.notificationText {
+  @include font-normal;
+  font-size: 1.4rem;
+  line-height: 2.2rem;
+}
+
+.notificationBlueStatus {
+  color: $blue-200;
+}
+
+.notificationRedStatus {
+  color: $red-400;
+}
+
+.notificationDate {
+  @include font-normal;
+  font-size: 1.2rem;
+  line-height: 1.6rem;
+  color: $gray-500;
+}
+
+@media (max-width: 767px) {
+  .notification {
+    position: initial;
+  }
+  .notificationWindow {
+    top: 0;
+    left: 0;
+    width: 100vw;
+    height: 100vh;
+    padding: 4rem 2rem;
+    border: 0;
+    border-radius: 0;
+    box-shadow: none;
+  }
+  .notificationClose {
+    display: block;
+  }
+  .notificationList {
+    height: 100%;
+    width: 100%;
+  }
+}
