+++ src/styles/_common.scss
@@ -1,2 +1,2 @@
 @import './_mixins';
-@import './_variables';
+@import './_color';
